<spike> Use consistent test names even when before hook fails
This is an exploration of the level of effort that would be involved in
addressing https://github.com/michaelleeallen/mocha-junit-reporter/issues/130.

Note that this change only explores the handling of `beforeEach` and
`beforeAll` hooks. It doesn't explore the handling of `afterEach` or
`afterAll` hooks which would likely also need some custom logic.

Also note that I've only tested this on Mocha 9.x currently.
mocha-junit-reporter's CI runs on a wide variety of Mocha versionse,
including some _really_ old versions, and I suspect this approach (in
its current form) won't work on some older versions of Mocha.